+++ src/js/action.js
@@ -41,7 +41,7 @@ var sendMessage = function (obj) {
   })
 }
 
-var setupMenus = function ($org, codelists, metadata) {
+var setupMenus = function ($org, codelists, metadata, version) {
   var $el = null
 
   $('#show-summary').on('click', function () {
@@ -122,7 +122,16 @@ var setupMenus = function ($org, codelists, metadata) {
     $el = $('#show-total-expenditure').on('click', function () {
       return false
     })
-    addNotProvidedPopup($el, 'Total expenditure', true)
+
+    var title = 'Total expenditure is not provided in this file.'
+    if (['1.01', '1.02', '1.03', '1.04', '1.05', '2.01'].indexOf(version) !== -1) {
+      title = 'This file uses IATI version ' + version + ', so total expenditure cannot be provided.'
+    }
+    $el.attr('title', title)
+      .attr('data-toggle', 'tooltip')
+      .attr('data-placement', 'bottom')
+      .parent()
+      .addClass('disabled')
   }
 
   // Documents menu item
@@ -292,9 +301,9 @@ $(function () {
         var $org = $orgs.first()
 
         $('#view-xml').attr('href', downloadUrl)
-        setupMenus($org, codelists, metadata)
-
         var version = $('iati-organisations', xml).attr('version')
+        setupMenus($org, codelists, metadata, version)
+
         var orgName = getNarrative('name', $org, version)
         $('#org-name').text(orgName)
         $('[data-toggle="tooltip"]').tooltip()