+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 ### Fixed
  - Fix alignment on second navbar ([#80](https://github.com/pwyf/iati-decipher/issues/80))
+ - Make it clear that `total-expenditure` is v2.02+ only ([#60](https://github.com/pwyf/iati-decipher/issues/60))
 
 ## [1.0.0] - 2018-10-30
 ### Changed